+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changes
 * [Developer/REST/UI]: Added announcements bell icon to main nav bar, added priority announcements modal at login, and refreshed user announcements page.
 * [UI]: Fixed bug where samples not rendering on project samples page when the config setting ngsarchive.linker.available is set to false.
 * [Developer]: Updated to Webpack v5 and Yarn v2.
+* [UI]: Updated create metadata template.
 * [UI/Developer]: Webpack does full bundle splitting to optimize loading time and thymeleaf injects the bundles onto template at compile time, see documentation for more details.
 
 20.09 to 21.01